refactor: move service package
